fix: modification of hitable struct to manage float damages => Laser deals expected damages

+++ Robot'em Up/Assets/Scripts/AI Behaviours/Base Enemy/EnemyBehaviour.cs
@@ -417,7 +417,7 @@ public class EnemyBehaviour : PawnController, IHitable
         }
     }
 
-    public virtual void OnHit(BallBehaviour _ball, Vector3 _impactVector, PawnController _thrower, int _damages, DamageSource _source, Vector3 _bumpModificators = default(Vector3))
+    public virtual void OnHit(BallBehaviour _ball, Vector3 _impactVector, PawnController _thrower, float _damages, DamageSource _source, Vector3 _bumpModificators = default(Vector3))
     {
         Vector3 i_normalizedImpactVector;
         LockManager.UnlockTarget(this.transform);
