+++ Robot'em Up/Assets/Scripts/NarrationElements/NarrativeInteractiveElements.cs
@@ -25,7 +25,7 @@ public class NarrativeInteractiveElements : MonoBehaviour, IHitable
 
     public AudioMixerGroup myAudioMixer;
 
-    public virtual void OnHit(BallBehaviour _ball, Vector3 _impactVector, PawnController _thrower, int _damages, DamageSource _source, Vector3 _bumpModificators = default)
+    public virtual void OnHit(BallBehaviour _ball, Vector3 _impactVector, PawnController _thrower, float _damages, DamageSource _source, Vector3 _bumpModificators = default)
     {
         if (!broken)
         {
